+++ Admin/view_sport_event.php
@@ -8,10 +8,10 @@
 
         session_start();
 
-        include "session_login.php";
+        include "../Shared/session_login.php";
         include "option_page.html";
-
         include_once "../Shared/connection.php";
+
         $query = "SELECT * from sport_event;";
         $sql_obj = mysqli_query($conn, $query);
         $count = mysqli_num_rows($sql_obj);
@@ -33,11 +33,12 @@
 
                     <h2>event:$ename
                     <img src = '../Images/$imgname' style = 'width: 350px;'></img>
-                    <h3>Stadium Name: $ename</h3>
-                    <h4>event start date:$sdate </h4>
-                    <h4>event time: $etime </h4>
+                    <h3>Stadium Name: $sname</h3>
+                    <h4>event start date:$edate </h4>
+                    <h4>event time: $stime </h4>
                     <p>event description:$desc</p>
-                    <button><a href='delete_event.php?ename=$ename'>Delete</button>
+                    <button onClick='return confirm(`Are you sure you want to cancel the event $ename?`)'><a href='delete_event.php?ename=$ename&edate=$edate'>Delete</a></button>
+                    <button><a href='edit_sport_event_html.php?ename=$ename&edate=$edate'>Edit</a></button>
                 </div>
             </div>
             ";
